+++ .github/scripts/bump-homebrew-tap.sh
@@ -82,6 +82,7 @@ class Bast < Formula
 
   def install
     bin.install "bast"
+    generate_completions_from_executable(bin/"bast", "completion", shells: [:bash, :zsh, :fish])
   end
 
   test do